# build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"postBookData": {
				"properties": {
					"title": {
						"type": "string"
					},
					"author": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"category": {
						"type": "number",
						"format": "double"
					},
					"language": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"title",
					"author",
					"description",
					"category",
					"language"
				],
				"type": "object",
				"additionalProperties": false
			},
			"putBookData": {
				"properties": {
					"bookId": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"author": {
						"type": "string"
					},
					"category": {
						"type": "number",
						"format": "double"
					},
					"language": {
						"type": "number",
						"format": "double"
					},
					"availability": {
						"type": "boolean"
					}
				},
				"required": [
					"bookId",
					"description",
					"id",
					"title",
					"author",
					"category",
					"language",
					"availability"
				],
				"type": "object",
				"additionalProperties": false
			},
			"postUserData": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"userName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"userName",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"putUserData": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"userName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"firstName",
					"userName",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "bookapplication",
		"version": "1.0.0",
		"description": "Application built on node ",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/getBook": {
			"get": {
				"operationId": "GetBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Book Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "bookId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/getAllBooks": {
			"get": {
				"operationId": "GetAllBooks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Book Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "pages",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limits",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/mostViewed": {
			"put": {
				"operationId": "MostViewed",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Book Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "bookId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/getBooks": {
			"get": {
				"operationId": "GetBooks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Book Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "bookId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/postBook": {
			"post": {
				"operationId": "PostBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Book Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/postBookData"
							}
						}
					}
				}
			}
		},
		"/api/putBook": {
			"put": {
				"operationId": "PutBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Book Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/putBookData"
							}
						}
					}
				}
			}
		},
		"/api/updateBook": {
			"put": {
				"operationId": "UpdateBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Book Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "bookId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "availability",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/api/deleteBook": {
			"delete": {
				"operationId": "DeleteBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Book Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "bookId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "availability",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/api/loginUser": {
			"post": {
				"operationId": "GetUserLogin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "For Getting Active user list",
				"summary": "Login User",
				"tags": [
					"User Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "password",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/userLogout": {
			"post": {
				"operationId": "UserLogout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "For Getting Active user list",
				"summary": "user Logout",
				"tags": [
					"User Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "password",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/getAllUser": {
			"get": {
				"operationId": "GetAllUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "For add user to the existing list",
				"summary": "get all user",
				"tags": [
					"User Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "pages",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limits",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/getUserById": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "For Getting Active user list",
				"summary": "User List",
				"tags": [
					"User Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "userid",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/addUser": {
			"post": {
				"operationId": "PostUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "For Getting Active user list",
				"summary": "Add User to List",
				"tags": [
					"User Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/postUserData"
							}
						}
					}
				}
			}
		},
		"/api/updateUser": {
			"put": {
				"operationId": "PutUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "For Getting Active user list",
				"summary": "Update User to List",
				"tags": [
					"User Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/putUserData"
							}
						}
					}
				}
			}
		},
		"/api/deleteUser": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "For Getting Active user list",
				"summary": "delate User from List",
				"tags": [
					"User Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "activeStatus",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}